refactor(settings): group the settings page and move operator tools to /admin
/settings was twelve peer accordions in one flat list, mixing everyday
preferences (superflex, TEP), power-user model tuning (two per-source
weight tables), personal data (watchlist, alerts) and operator surfaces
— all at the same visual weight, with nothing saying which was which.
Finding anything meant opening sections until you hit it.

Same sections, same order within each, now under three headings:

  League & scoring          — what your league is; changes what the
                              numbers mean
  How values are calculated — the valuation engine's knobs
  Alerts & lists            — notifications, watchlist, custom alerts

Two sections left the page entirely:

  * Data & Admin (server status + manual scrape trigger) and Guest
    access (minting and revoking credentials) moved to /admin. Both
    were operator actions filed under preferences, on a page every
    signed-in user opens to change their TEP setting, and their
    endpoints are admin-gated anyway. They are now
    components/admin/{ServerStatusPanel,GuestPassPanel}.jsx — the move
    is verbatim, only the export changed. An admin-only line in the
    settings footer says where they went.

  * Pick Settings contained no controls. It was one paragraph
    explaining that a setting had been REMOVED — a section that existed
    to describe its own absence.

Also fixes the accordion's a11y: the hand-rolled disclosure carried no
aria-expanded and no aria-controls, so assistive tech announced twelve
identical buttons with no state and no relationship to the panels they
toggle. The +/- glyph is now aria-hidden, since it was the only
indication of state and it was decorative.

/settings drops from 2003 to ~1530 lines.

One backend test needed updating: test_finder_routes_stay_distinct
pinned /arbitrage's nav label as the exact string "Arbitrage Finder".
Its invariant is that the arbitrage vocabulary must not land on
/finder, which computes none — that still holds and is now asserted on
the word rather than the exact label, because the nav says "Arbitrage"
to match the page's own <h1> (they disagreed before, which was its own
small part of the confusion the test exists to prevent).

Tests: 5323 backend + 366 subtests (was 5306), 1425 frontend.
